Add default input, simplify tests

+++ impact/parsers.py
@@ -1600,7 +1600,7 @@ def parse_impact_particles(filePath,
     """
     
     dtype={'names': names,
-           'formats': 6*[np.float]}
+           'formats': 6*[float]}
     pdat = np.loadtxt(filePath, skiprows=skiprows, dtype=dtype,
                      ndmin=1) # to make sure that 1 particle is parsed the same as many.
